// src/app/modules/matenimiento/models/proveedor/requestPedidoProveedor.model.ts
export class RequestComprobanteProveedor {
    fechaEntrega: string="";
    fechaPedido: string="";
    costeTransporte: number=0;
    nombreProveedor :string=""
    idComprobante: number=0;
    tipoComprobante: string="";
    codigoComprobante: string="";
    idProvedor: number=0;
    montoTotal: number=0 ;
    cantidadTotal: number=0 ;
    subTotal: number=0 ;
}